requirements.txt
jsonschema==2.3.0
ujs-safefile==0.1.0
